# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: preorder
    description: Guest pre-order flow (menu publish, orders)
  - name: feedback
    description: Guest feedback / multi-criteria ratings
  - name: rollup
    description: Server-side multi-location aggregations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /preorder/menu/{locationCode}:
    get:
      operationId: getPublishedMenu
      tags: [preorder]
      summary: Public menu lookup by short location code (used by guest QR app).
      parameters:
        - in: path
          name: locationCode
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Current published menu for the location
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublishedMenu"
        "404":
          description: Location not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /preorder/menu/publish:
    post:
      operationId: publishMenu
      tags: [preorder]
      summary: Staff pushes the current menu for a location (requires auth).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PublishMenuBody"
      responses:
        "200":
          description: Published
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublishedMenu"

  /preorder/orders:
    post:
      operationId: createGuestOrder
      tags: [preorder]
      summary: Guest creates an order (no auth, returns access token for status polling).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderBody"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GuestOrder"
        "400":
          description: Invalid order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /preorder/orders/{id}:
    get:
      operationId: getGuestOrder
      tags: [preorder]
      summary: Public order status lookup using id + access token.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
        - in: query
          name: token
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GuestOrder"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /preorder/orders/{id}/status:
    patch:
      operationId: updateOrderStatus
      tags: [preorder]
      summary: Staff transitions an order's status (requires auth).
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderStatusBody"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GuestOrder"

  /preorder/feedback:
    post:
      operationId: createFeedback
      tags: [feedback]
      summary: Guest submits multi-criteria feedback for a location (no auth).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFeedbackBody"
      responses:
        "201":
          description: Feedback recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GuestFeedback"
        "400":
          description: Invalid feedback
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /preorder/staff/feedback:
    get:
      operationId: listStaffFeedback
      tags: [feedback]
      summary: Staff lists recent feedback for a location with aggregate (requires auth).
      parameters:
        - in: query
          name: locationCode
          required: true
          schema:
            type: string
        - in: query
          name: days
          required: false
          schema:
            type: integer
            default: 30
      responses:
        "200":
          description: Feedback list with aggregate ratings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedbackSummary"

  /rollup/locations:
    get:
      operationId: getLocationRollup
      tags: [rollup]
      summary: Aggregate orders + feedback across all of the caller's locations (requires auth).
      parameters:
        - in: query
          name: days
          required: false
          schema:
            type: integer
            default: 7
      responses:
        "200":
          description: Per-location rollup
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LocationRollup"

  /preorder/staff/orders:
    get:
      operationId: listStaffOrders
      tags: [preorder]
      summary: Staff lists pending/active orders for a location (requires auth).
      parameters:
        - in: query
          name: locationCode
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Order list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GuestOrder"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]

    PreorderDish:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        type:
          type: ["string", "null"]
        allergens:
          type: array
          items:
            type: string
        imageUrl:
          type: ["string", "null"]
        kcal:
          type: ["number", "null"]
        proteinG:
          type: ["number", "null"]
        carbsG:
          type: ["number", "null"]
        fatG:
          type: ["number", "null"]
        dge:
          type: ["string", "null"]
          enum: [green, amber, red, null]
        co2eG:
          type: ["number", "null"]
      required: [id, name, price, allergens]

    FeedbackRatings:
      type: object
      properties:
        foodQuality:
          type: integer
          minimum: 1
          maximum: 5
        service:
          type: integer
          minimum: 1
          maximum: 5
        variety:
          type: integer
          minimum: 1
          maximum: 5
        value:
          type: integer
          minimum: 1
          maximum: 5
        cleanliness:
          type: integer
          minimum: 1
          maximum: 5
        ambience:
          type: integer
          minimum: 1
          maximum: 5
      required: [foodQuality, service, variety, value, cleanliness, ambience]

    GuestFeedback:
      type: object
      properties:
        id:
          type: string
        locationCode:
          type: string
        ratings:
          $ref: "#/components/schemas/FeedbackRatings"
        overall:
          type: number
        comment:
          type: ["string", "null"]
        guestName:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, locationCode, ratings, overall, createdAt]

    CreateFeedbackBody:
      type: object
      properties:
        locationCode:
          type: string
          minLength: 1
          maxLength: 32
        ratings:
          $ref: "#/components/schemas/FeedbackRatings"
        comment:
          type: ["string", "null"]
          maxLength: 500
        guestName:
          type: ["string", "null"]
          maxLength: 60
      required: [locationCode, ratings]

    FeedbackAggregate:
      type: object
      properties:
        count:
          type: integer
        avgFoodQuality:
          type: number
        avgService:
          type: number
        avgVariety:
          type: number
        avgValue:
          type: number
        avgCleanliness:
          type: number
        avgAmbience:
          type: number
        avgOverall:
          type: number
      required:
        [
          count,
          avgFoodQuality,
          avgService,
          avgVariety,
          avgValue,
          avgCleanliness,
          avgAmbience,
          avgOverall,
        ]

    FeedbackSummary:
      type: object
      properties:
        aggregate:
          $ref: "#/components/schemas/FeedbackAggregate"
        recent:
          type: array
          items:
            $ref: "#/components/schemas/GuestFeedback"
      required: [aggregate, recent]

    LocationRollup:
      type: object
      properties:
        locationCode:
          type: string
        locationName:
          type: string
        currency:
          type: string
        ordersCount:
          type: integer
        revenue:
          type: number
        avgTicket:
          type: number
        statusBreakdown:
          type: object
          additionalProperties:
            type: integer
        feedback:
          $ref: "#/components/schemas/FeedbackAggregate"
        publishedAt:
          type: ["string", "null"]
      required:
        [locationCode, locationName, currency, ordersCount, revenue, avgTicket, statusBreakdown, feedback]

    PublishedMenu:
      type: object
      properties:
        locationCode:
          type: string
        locationName:
          type: string
        currency:
          type: string
        dishes:
          type: array
          items:
            $ref: "#/components/schemas/PreorderDish"
        publishedAt:
          type: string
      required: [locationCode, locationName, currency, dishes, publishedAt]

    PublishMenuBody:
      type: object
      properties:
        locationCode:
          type: string
        locationName:
          type: string
        currency:
          type: string
        dishes:
          type: array
          items:
            $ref: "#/components/schemas/PreorderDish"
      required: [locationCode, locationName, dishes]

    OrderItem:
      type: object
      properties:
        dishId:
          type: string
        name:
          type: string
        qty:
          type: integer
        price:
          type: number
      required: [dishId, name, qty, price]

    GuestOrderStatus:
      type: string
      enum: [new, accepted, preparing, ready, served, cancelled]

    GuestOrder:
      type: object
      properties:
        id:
          type: string
        locationCode:
          type: string
        guestName:
          type: string
        guestNote:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        total:
          type: number
        currency:
          type: string
        status:
          $ref: "#/components/schemas/GuestOrderStatus"
        accessToken:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, locationCode, guestName, items, total, currency, status, createdAt, updatedAt]

    CreateOrderBody:
      type: object
      properties:
        locationCode:
          type: string
        guestName:
          type: string
        guestNote:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
      required: [locationCode, guestName, items]

    UpdateOrderStatusBody:
      type: object
      properties:
        status:
          $ref: "#/components/schemas/GuestOrderStatus"
      required: [status]