# lib/api-zod/src/generated/types/createFeedbackBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { FeedbackRatings } from "./feedbackRatings";

export interface CreateFeedbackBody {
  /**
   * @minLength 1
   * @maxLength 32
   */
  locationCode: string;
  ratings: FeedbackRatings;
  /**
   * @maxLength 500
   * @nullable
   */
  comment?: string | null;
  /**
   * @maxLength 60
   * @nullable
   */
  guestName?: string | null;
}