# lib/api-zod/src/generated/types/feedbackAggregate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface FeedbackAggregate {
  count: number;
  avgFoodQuality: number;
  avgService: number;
  avgVariety: number;
  avgValue: number;
  avgCleanliness: number;
  avgAmbience: number;
  avgOverall: number;
}