# lib/api-zod/src/generated/types/feedbackRatings.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface FeedbackRatings {
  /**
   * @minimum 1
   * @maximum 5
   */
  foodQuality: number;
  /**
   * @minimum 1
   * @maximum 5
   */
  service: number;
  /**
   * @minimum 1
   * @maximum 5
   */
  variety: number;
  /**
   * @minimum 1
   * @maximum 5
   */
  value: number;
  /**
   * @minimum 1
   * @maximum 5
   */
  cleanliness: number;
  /**
   * @minimum 1
   * @maximum 5
   */
  ambience: number;
}