# lib/api-zod/src/generated/types/feedbackSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { FeedbackAggregate } from "./feedbackAggregate";
import type { GuestFeedback } from "./guestFeedback";

export interface FeedbackSummary {
  aggregate: FeedbackAggregate;
  recent: GuestFeedback[];
}