# lib/api-zod/src/generated/types/guestFeedback.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { FeedbackRatings } from "./feedbackRatings";

export interface GuestFeedback {
  id: string;
  locationCode: string;
  ratings: FeedbackRatings;
  overall: number;
  /** @nullable */
  comment?: string | null;
  /** @nullable */
  guestName?: string | null;
  createdAt: string;
}