# lib/api-zod/src/generated/types/locationRollup.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { FeedbackAggregate } from "./feedbackAggregate";
import type { LocationRollupStatusBreakdown } from "./locationRollupStatusBreakdown";

export interface LocationRollup {
  locationCode: string;
  locationName: string;
  currency: string;
  ordersCount: number;
  revenue: number;
  avgTicket: number;
  statusBreakdown: LocationRollupStatusBreakdown;
  feedback: FeedbackAggregate;
  /** @nullable */
  publishedAt?: string | null;
}