# lib/api-zod/src/generated/types/preorderDish.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PreorderDishDge } from "./preorderDishDge";

export interface PreorderDish {
  id: string;
  name: string;
  /** @nullable */
  description?: string | null;
  price: number;
  /** @nullable */
  type?: string | null;
  allergens: string[];
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  kcal?: number | null;
  /** @nullable */
  proteinG?: number | null;
  /** @nullable */
  carbsG?: number | null;
  /** @nullable */
  fatG?: number | null;
  /** @nullable */
  dge?: PreorderDishDge;
  /** @nullable */
  co2eG?: number | null;
}